Add files via upload
Apriori is an algorithm, which uses an iterative approach to find frequent itemsets in a transactional database and generate association rules from these frequent itemsets. It is based on an observation that an itemset is frequent only if all its non-empty subsets are frequent. Each kthiteration of Apriori generates frequent patterns of length k.